requirements.txt
django==1.10.5
requests
Pillow
bs4
lxml
gunicorn
djangorestframework
Pygments